requirements.txt
torch
torchvision
numpy
matplotlib
scikit-learn
tqdm
pillow
pandas
pyyaml